Refactor signup dom test extracting process to be reused
